FINALLY A WORKING PROTOTYPE

+++ src/math/Vector.java
@@ -1,28 +1,28 @@
 package math;
 
 public class Vector {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public Vector(int x, int y)
+    public Vector(float x, float y)
     {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
